WiP breaking changes: refactoring, split attachment (category and access), rename packages
