+++ app/crew/grails-app/domain/attachment/DocumentCategory.groovy
@@ -1,0 +1,19 @@
+package attachment
+
+import attachment.config.DocumentCategoryEnum
+import grails.compiler.GrailsCompileStatic
+import taack.ast.annotation.TaackFieldEnum
+
+@TaackFieldEnum
+@GrailsCompileStatic
+class DocumentCategory {
+
+    DocumentCategoryEnum category = DocumentCategoryEnum.OTHER
+    Set<Term> tags
+
+    static constraints = {}
+
+    static hasMany = [
+            tags: Term
+    ]
+}
