+++ app/crew/grails-app/domain/attachment/TaackDocument.groovy
@@ -1,0 +1,18 @@
+package attachment
+
+import crew.User
+import grails.compiler.GrailsCompileStatic
+import taack.ast.annotation.TaackFieldEnum
+
+@TaackFieldEnum
+@GrailsCompileStatic
+abstract class TaackDocument {
+    User userCreated
+    Date dateCreated
+    User userUpdated
+    Date lastUpdated
+
+    DocumentAccess documentAccess
+    DocumentCategory documentCategory
+
+}
