+++ app/crew/grails-app/domain/crew/PersistentLogin.groovy
@@ -1,4 +1,4 @@
-package org.taack
+package crew
 
 import groovy.transform.EqualsAndHashCode
 import groovy.transform.ToString
