+++ app/crew/grails-app/domain/crew/UserRole.groovy
@@ -1,4 +1,5 @@
-package org.taack
+package crew
+
 
 import grails.gorm.DetachedCriteria
 import groovy.transform.ToString
