+++ app/crew/grails-app/services/crew/CrewSearchService.groovy
@@ -4,7 +4,6 @@ package crew
 import grails.compiler.GrailsCompileStatic
 import org.codehaus.groovy.runtime.MethodClosure
 import org.grails.datastore.gorm.GormEntity
-import org.taack.User
 import taack.domain.TaackSearchService
 import taack.solr.SolrSpecifier
 import taack.solr.SolrFieldType
