+++ app/crew/src/main/groovy/attachment/SshAttachmentFolder.groovy
@@ -1,10 +1,9 @@
 package attachment
 
+import crew.User
 import grails.util.Pair
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
-import org.taack.Attachment
-import org.taack.User
 import taack.domain.TaackAttachmentService
 import taack.ssh.vfs.FileCallback
 import taack.ssh.vfs.FileTree
