+++ app/crew/src/main/groovy/attachment/config/TermGroupConfig.groovy
@@ -1,4 +1,4 @@
-package app.config
+package attachment.config
 
 import groovy.transform.CompileStatic
