+++ app/crew/src/main/groovy/org/taack/IAttachmentConverter.groovy
@@ -1,5 +1,7 @@
 package org.taack
 
+import attachment.Attachment
+
 interface IAttachmentConverter {
     Map<String, List<String>> getSupportedExtensionConversions()
     File convertTo(Attachment attachment, String extensionTo)
