+++ app/crew/src/main/groovy/org/taack/IAttachmentPreviewConverter.groovy
@@ -1,5 +1,7 @@
 package org.taack
 
+import attachment.Attachment
+
 interface IAttachmentPreviewConverter {
     List<String> getPreviewManagedExtensions()
     void createWebpPreview(Attachment attachment, String previewPath)
