+++ app/crew/src/main/groovy/org/taack/IAttachmentShowIFrame.groovy
@@ -1,5 +1,7 @@
 package org.taack
 
+import attachment.Attachment
+
 interface IAttachmentShowIFrame {
     List<String> getShowIFrameManagedExtensions()
     String createShowIFrame(Attachment attachment)
